Adding lib directory for ruby scripts
